Use QueryValidation in message exports
