+++ full-backend-tests/src/test/java/org/graylog/plugins/views/MessagesResourceIT.java
@@ -17,17 +17,20 @@
 package org.graylog.plugins.views;
 
 import io.restassured.response.Response;
+import io.restassured.response.ValidatableResponse;
 import io.restassured.specification.RequestSpecification;
+import org.assertj.core.api.Assertions;
 import org.graylog.testing.completebackend.GraylogBackend;
-import org.graylog.testing.containermatrix.MongodbServer;
-import org.graylog.testing.containermatrix.SearchServer;
 import org.graylog.testing.containermatrix.annotations.ContainerMatrixTest;
 import org.graylog.testing.containermatrix.annotations.ContainerMatrixTestsConfiguration;
+import org.hamcrest.Matchers;
+import org.junit.jupiter.api.BeforeAll;
 
 import java.util.Arrays;
 
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.core.IsEqual.equalTo;
 
 @ContainerMatrixTestsConfiguration
 public class MessagesResourceIT {
@@ -39,8 +42,26 @@ public class MessagesResourceIT {
         this.requestSpec = requestSpec;
     }
 
+    @BeforeAll
+    public void importMessages() {
+        sut.importElasticsearchFixture("messages-for-export.json", MessagesResourceIT.class);
+    }
+
     @ContainerMatrixTest
-    void canDownloadCsv() {
+    void testInvalidQuery() {
+        String allMessagesTimeRange = "{\"query_string\":\"foo:\", \"timerange\": {\"type\": \"absolute\", \"from\": \"2015-01-01T00:00:00\", \"to\": \"2015-01-01T23:59:59\"}}";
+        given()
+                .spec(requestSpec)
+                .accept("text/csv")
+                .body(allMessagesTimeRange)
+                .post("/views/search/messages")
+                .then()
+                .statusCode(400).contentType("application/json")
+                .assertThat().body("message", Matchers.startsWith("Request validation failed"));
+    }
+
+    @ContainerMatrixTest
+    void testInvalidQueryResponse() {
         sut.importElasticsearchFixture("messages-for-export.json", MessagesResourceIT.class);
 
         String allMessagesTimeRange = "{\"timerange\": {\"type\": \"absolute\", \"from\": \"2015-01-01T00:00:00\", \"to\": \"2015-01-01T23:59:59\"}}";
