+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/rest/MessagesResource.java
@@ -42,6 +42,11 @@ import org.graylog.plugins.views.search.export.SearchExportJob;
 import org.graylog.plugins.views.search.export.SearchTypeExportJob;
 import org.graylog.plugins.views.search.export.SimpleMessageChunk;
 import org.graylog.plugins.views.search.permissions.SearchUser;
+import org.graylog.plugins.views.search.validation.QueryValidationService;
+import org.graylog.plugins.views.search.validation.ValidationMessage;
+import org.graylog.plugins.views.search.validation.ValidationRequest;
+import org.graylog.plugins.views.search.validation.ValidationResponse;
+import org.graylog.plugins.views.search.validation.ValidationStatus;
 import org.graylog2.audit.jersey.NoAuditEvent;
 import org.graylog2.plugin.rest.PluginRestResource;
 import org.graylog2.rest.MoreMediaTypes;
@@ -50,6 +55,7 @@ import org.joda.time.DateTimeZone;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
+import javax.ws.rs.BadRequestException;
 import javax.ws.rs.GET;
 import javax.ws.rs.NotFoundException;
 import javax.ws.rs.POST;
@@ -59,6 +65,7 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import java.io.UnsupportedEncodingException;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
@@ -73,6 +80,7 @@ public class MessagesResource extends RestResource implements PluginRestResource
     private final PermittedStreams permittedStreams;
     private final ObjectMapper objectMapper;
     private final ExportJobService exportJobService;
+    private final QueryValidationService queryValidationService;
 
     //allow mocking
     Function<Consumer<Consumer<SimpleMessageChunk>>, ChunkedOutput<SimpleMessageChunk>> asyncRunner = ChunkedRunner::runAsync;
@@ -87,13 +95,14 @@ public class MessagesResource extends RestResource implements PluginRestResource
             PermittedStreams permittedStreams,
             ObjectMapper objectMapper,
             @SuppressWarnings("UnstableApiUsage") EventBus eventBus,
-            ExportJobService exportJobService) {
+            ExportJobService exportJobService, QueryValidationService queryValidationService) {
         this.commandFactory = commandFactory;
         this.searchDomain = searchDomain;
         this.executionGuard = executionGuard;
         this.permittedStreams = permittedStreams;
         this.objectMapper = objectMapper;
         this.exportJobService = exportJobService;
+        this.queryValidationService = queryValidationService;
         this.messagesExporterFactory = context -> new AuditingMessagesExporter(context, eventBus, exporter);
     }
 
@@ -106,7 +115,19 @@ public class MessagesResource extends RestResource implements PluginRestResource
     @NoAuditEvent("Has custom audit events")
     public ChunkedOutput<SimpleMessageChunk> retrieve(@ApiParam @Valid MessagesRequest rawrequest,
                                                       @Context SearchUser searchUser) {
+
         final MessagesRequest request = fillInIfNecessary(rawrequest, searchUser);
+
+        final ValidationRequest.Builder validationReq = ValidationRequest.builder();
+        Optional.ofNullable(rawrequest.queryString()).ifPresent(validationReq::query);
+        Optional.ofNullable(rawrequest.timeRange()).ifPresent(validationReq::timerange);
+        Optional.ofNullable(rawrequest.streams()).ifPresent(validationReq::streams);
+        final ValidationResponse validationResponse = queryValidationService.validate(validationReq.build());
+        if (validationResponse.status().equals(ValidationStatus.ERROR)) {
+            validationResponse.explanations().stream().findFirst().map(ValidationMessage::errorMessage).ifPresent(message -> {
+                throw new BadRequestException("Request validation failed: " + message);
+            });
+        }
 
         executionGuard.checkUserIsPermittedToSeeStreams(request.streams(), searchUser::canReadStream);
